IC-2114 Wrong pUniswap icon and name in PDEX screen, change coin service url
