+++ src/components/Icons/icon.uni.js
@@ -1,60 +1,21 @@
 import React from 'react';
-import Svg, { Path } from 'react-native-svg';
+import { StyleSheet, ImageProps } from 'react-native';
+import { Image } from '@src/components/core';
+import { COLORS } from '@src/styles';
+import uniSrcIcon from '@src/assets/images/new-icons/uni.png';
 
-const UniIcon = (props) => (
-  <Svg
-    width={50}
-    height={50}
-    viewBox="0 0 50 50"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    {...props}
-  >
-    <Path
-      d="M49.252 31.0476C45.9127 44.4406 32.3462 52.5913 18.9501 49.2515C5.55965 45.9126 -2.59198 32.3468 0.748798 18.9549C4.08664 5.56048 17.6531 -2.59085 31.045 0.748085C44.4402 4.08702 52.5913 17.6543 49.2516 31.0479L49.2519 31.0476H49.252Z"
-      fill="#FEF3F9"
-    />
-    <Path
-      d="M21.0547 15.4406C20.7162 15.3922 20.7001 15.3761 20.8613 15.36C21.1675 15.3116 21.8766 15.3761 22.3762 15.4889C23.5366 15.7629 24.5842 16.472 25.6962 17.713L25.9863 18.0515L26.4054 17.987C28.1943 17.6969 30.0316 17.9225 31.5627 18.6317C31.9817 18.8251 32.6425 19.2118 32.7231 19.3085C32.7553 19.3408 32.8036 19.5503 32.8359 19.7598C32.9487 20.5173 32.9003 21.0814 32.6586 21.5165C32.5297 21.7583 32.5297 21.8227 32.6102 22.0322C32.6747 22.1934 32.8681 22.3062 33.0454 22.3062C33.4322 22.3062 33.8351 21.6938 34.0285 20.8396L34.1091 20.5012L34.2541 20.6623C35.0761 21.581 35.7207 22.8542 35.8174 23.7567L35.8497 23.9985L35.7046 23.789C35.4629 23.4183 35.2372 23.1765 34.931 22.967C34.3831 22.5963 33.8029 22.4835 32.2718 22.4029C30.8858 22.3223 30.0961 22.2095 29.3225 21.9517C28.0009 21.5165 27.324 20.9524 25.7607 18.8734C25.0677 17.9548 24.6325 17.4551 24.1974 17.0361C23.2465 16.1175 22.2956 15.634 21.0547 15.4406Z"
-      fill="#FF007A"
-    />
-    <Path
-      d="M33.0614 17.4881C33.0937 16.8757 33.1743 16.4728 33.3515 16.1021C33.416 15.9571 33.4805 15.8281 33.4966 15.8281C33.5127 15.8281 33.4805 15.9409 33.4321 16.0699C33.3032 16.4244 33.2871 16.9241 33.3677 17.4881C33.4805 18.2134 33.5288 18.3101 34.3024 19.0998C34.657 19.4705 35.076 19.9378 35.2372 20.1312L35.5112 20.4858L35.2372 20.2279C34.8987 19.9056 34.1251 19.2932 33.9478 19.2126C33.835 19.1481 33.8189 19.1481 33.7383 19.2287C33.6739 19.2932 33.6578 19.3899 33.6578 19.8573C33.6416 20.5825 33.5449 21.0338 33.3032 21.5012C33.1743 21.7429 33.1581 21.6946 33.271 21.4206C33.3515 21.2111 33.3677 21.1144 33.3677 20.4213C33.3677 19.0192 33.2065 18.6808 32.2234 18.1167C31.9816 17.9716 31.5626 17.7621 31.3209 17.6493C31.063 17.5365 30.8696 17.4398 30.8857 17.4398C30.9179 17.4075 31.8688 17.6815 32.2395 17.8427C32.8036 18.0683 32.9003 18.0844 32.9647 18.0683C33.0131 18.02 33.0453 17.891 33.0614 17.4881Z"
-      fill="#FF007A"
-    />
-    <Path
-      d="M21.7153 19.8894C21.0384 18.9547 20.6032 17.5042 20.6999 16.4244L20.7322 16.0859L20.8933 16.1182C21.1834 16.1665 21.683 16.3599 21.9248 16.505C22.5695 16.8918 22.8596 17.4236 23.1335 18.7452C23.2141 19.132 23.3269 19.5832 23.3753 19.7283C23.4559 19.97 23.7621 20.5341 24.0199 20.8887C24.1972 21.1465 24.0844 21.2755 23.6815 21.2432C23.0691 21.1788 22.2471 20.6147 21.7153 19.8894Z"
-      fill="#FF007A"
-    />
-    <Path
-      d="M32.2392 26.8999C29.0481 25.6106 27.9199 24.4985 27.9199 22.6129C27.9199 22.3389 27.936 22.1133 27.936 22.1133C27.9522 22.1133 28.065 22.21 28.21 22.3228C28.8547 22.8385 29.5799 23.0642 31.5945 23.3543C32.771 23.5315 33.4479 23.6605 34.0603 23.87C36.0104 24.5146 37.2192 25.8362 37.5093 27.6251C37.5899 28.1409 37.5415 29.124 37.4126 29.6397C37.2998 30.0426 36.9774 30.784 36.8968 30.8001C36.8807 30.8001 36.8485 30.7195 36.8485 30.5906C36.8163 29.9137 36.4778 29.269 35.9137 28.7694C35.2368 28.1892 34.3665 27.7541 32.2392 26.8999Z"
-      fill="#FF007A"
-    />
-    <Path
-      d="M29.9829 27.4323C29.9506 27.1906 29.87 26.8844 29.8217 26.7554L29.7411 26.5137L29.8862 26.691C30.0957 26.9327 30.2568 27.2228 30.4019 27.6257C30.5147 27.9319 30.5147 28.0286 30.5147 28.5282C30.5147 29.0117 30.4986 29.1246 30.4019 29.3985C30.2407 29.8337 30.0473 30.1399 29.725 30.4783C29.1448 31.0747 28.3873 31.397 27.3075 31.542C27.1141 31.5582 26.5661 31.6065 26.0826 31.6387C24.8739 31.7032 24.0681 31.8321 23.3428 32.09C23.2461 32.1222 23.1494 32.1545 23.1333 32.1383C23.1011 32.1061 23.6007 31.816 24.0036 31.6226C24.5677 31.3486 25.1479 31.2036 26.4211 30.978C27.0496 30.8813 27.6943 30.7523 27.8555 30.6879C29.451 30.1882 30.2407 28.9473 29.9829 27.4323Z"
-      fill="#FF007A"
-    />
-    <Path
-      d="M31.4496 30.0269C31.0305 29.1083 30.9338 28.238 31.1595 27.416C31.1917 27.3355 31.2239 27.2549 31.2562 27.2549C31.2884 27.2549 31.3851 27.3032 31.4818 27.3677C31.6752 27.4966 32.0781 27.7223 33.1096 28.2863C34.415 28.9955 35.1564 29.5434 35.6721 30.172C36.1234 30.7199 36.3974 31.3485 36.5263 32.1221C36.6069 32.5572 36.5585 33.6048 36.4457 34.04C36.0911 35.4099 35.2853 36.5058 34.1088 37.1343C33.9315 37.231 33.7865 37.2955 33.7704 37.2955C33.7542 37.2955 33.8187 37.1343 33.9154 36.9409C34.3022 36.119 34.3506 35.3293 34.0605 34.4429C33.8832 33.8949 33.5125 33.2341 32.7711 32.1221C31.8847 30.8328 31.6752 30.4943 31.4496 30.0269Z"
-      fill="#FF007A"
-    />
-    <Path
-      d="M19.4429 34.958C20.6356 33.9588 22.1022 33.2497 23.456 33.024C24.0362 32.9273 25.0032 32.9596 25.535 33.1046C26.3892 33.3302 27.1628 33.8137 27.5657 34.41C27.9525 34.9902 28.1298 35.4899 28.307 36.6019C28.3715 37.037 28.4521 37.4883 28.4682 37.585C28.5971 38.1652 28.855 38.6165 29.1773 38.8582C29.677 39.2289 30.5472 39.245 31.4014 38.9227C31.5465 38.8743 31.6754 38.826 31.6754 38.8421C31.7076 38.8743 31.2725 39.1644 30.9824 39.3095C30.5795 39.519 30.2571 39.5835 29.822 39.5835C29.0484 39.5835 28.3876 39.1805 27.8558 38.3747C27.743 38.2136 27.5173 37.7462 27.3239 37.311C26.7599 36.0056 26.4698 35.6188 25.809 35.1836C25.2288 34.813 24.4874 34.7324 23.9233 35.0064C23.182 35.3609 22.9886 36.3118 23.5043 36.892C23.7138 37.1337 24.1006 37.3271 24.423 37.3755C25.0193 37.4561 25.535 36.9887 25.535 36.3924C25.535 36.0056 25.3899 35.78 25.0032 35.6027C24.4874 35.377 23.9233 35.6349 23.9395 36.1345C23.9395 36.344 24.0362 36.473 24.2457 36.5697C24.3746 36.6341 24.3746 36.6341 24.2779 36.618C23.8105 36.5213 23.6977 35.9411 24.0684 35.5704C24.5197 35.1192 25.4705 35.3126 25.7929 35.9411C25.9218 36.199 25.9379 36.7147 25.8251 37.037C25.5511 37.7462 24.7775 38.1169 23.9878 37.9073C23.456 37.7623 23.2303 37.6172 22.5857 36.9565C21.4575 35.7961 21.0224 35.5704 19.4107 35.3287L19.1045 35.2803L19.4429 34.958Z"
-      fill="#FF007A"
-    />
-    <Path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M10.9657 9.027C14.7208 13.588 20.5067 20.6793 20.7968 21.0661C21.0385 21.3884 20.9418 21.6946 20.5389 21.9202C20.3133 22.0492 19.8459 22.1781 19.6203 22.1781C19.3624 22.1781 19.0562 22.0492 18.8467 21.8397C18.7016 21.6946 18.0731 20.776 16.6548 18.568C15.575 16.8758 14.6564 15.4736 14.6402 15.4575C14.5758 15.4253 14.5758 15.4253 16.542 18.9387C17.783 21.1466 18.1859 21.9364 18.1859 22.0331C18.1859 22.2426 18.1214 22.3554 17.8636 22.6455C17.4284 23.129 17.235 23.6769 17.09 24.8212C16.9288 26.0944 16.4937 26.997 15.2527 28.528C14.5274 29.4306 14.4146 29.5917 14.2373 29.9624C14.0117 30.4137 13.9472 30.6715 13.915 31.2517C13.8828 31.8642 13.9472 32.251 14.1245 32.8311C14.2857 33.3469 14.463 33.6853 14.8981 34.3461C15.2688 34.9263 15.4944 35.3614 15.4944 35.5226C15.4944 35.6515 15.5267 35.6515 16.1069 35.5226C17.4929 35.2003 18.6372 34.6523 19.2657 33.9754C19.6525 33.5564 19.7492 33.3308 19.7492 32.7506C19.7492 32.3799 19.7331 32.2993 19.6364 32.0737C19.4752 31.7191 19.169 31.429 18.5082 30.9777C17.6379 30.3814 17.2672 29.8979 17.1705 29.2533C17.09 28.7053 17.1867 28.3346 17.6702 27.3193C18.1698 26.2717 18.2987 25.8366 18.3793 24.7729C18.4276 24.096 18.5082 23.822 18.7016 23.6125C18.9111 23.3868 19.0884 23.3063 19.588 23.2418C20.41 23.129 20.9418 22.9195 21.3609 22.5166C21.7315 22.1781 21.8927 21.8397 21.9088 21.34L21.9249 20.9694L21.7154 20.7437C20.9579 19.8734 10.4661 8.33398 10.4177 8.33398C10.4016 8.33398 10.6595 8.6402 10.9657 9.027ZM15.9134 31.9448C16.0907 31.6385 15.994 31.2517 15.7039 31.0583C15.4299 30.8811 15.0109 30.9616 15.0109 31.2034C15.0109 31.2679 15.0431 31.3323 15.1398 31.3646C15.2849 31.4451 15.301 31.5257 15.1882 31.703C15.0754 31.8803 15.0754 32.0415 15.2204 32.1543C15.4461 32.3315 15.7523 32.2348 15.9134 31.9448Z"
-      fill="#FF007A"
-    />
-    <Path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M22.4407 23.4662C22.054 23.579 21.6833 23.998 21.5705 24.417C21.506 24.6749 21.5382 25.1423 21.651 25.2873C21.8283 25.513 21.9895 25.5774 22.4407 25.5774C23.3272 25.5774 24.0846 25.1906 24.1652 24.7232C24.2458 24.3364 23.9074 23.8046 23.44 23.5629C23.1982 23.4339 22.6986 23.3856 22.4407 23.4662ZM23.4722 24.2721C23.6011 24.0787 23.5528 23.8692 23.311 23.7241C22.8759 23.4501 22.2151 23.6758 22.2151 24.0948C22.2151 24.3043 22.5535 24.53 22.8759 24.53C23.0854 24.53 23.3755 24.401 23.4722 24.2721Z"
-      fill="#FF007A"
-    />
-  </Svg>
-);
+const styled = StyleSheet.create({
+  icon: {
+    width: 20,
+    height: 20,
+  },
+});
+
+const UniIcon = (props: ImageProps) => {
+  const { style, ...rest } = props;
+  return <Image source={uniSrcIcon} style={[styled.icon, style]} {...rest} />;
+};
+
+UniIcon.propTypes = {};
 
 export default UniIcon;
