+++ src/screens/PDexV3/features/PrivacyApps/PrivacyApps.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import { StyleSheet } from 'react-native';
 import { withLayout_2 } from '@src/components/Layout';
-// import { ScrollViewBorder } from '@components/core';
 import { View } from '@src/components/core';
-import { PancakeIcon2, UniIcon } from '@src/components/Icons';
+import { PancakeIcon2, UniIcon2 } from '@src/components/Icons';
 import { useDispatch, useSelector } from 'react-redux';
 import { FONT } from '@src/styles';
 import { KEYS_PLATFORMS_SUPPORTED } from '@screens/PDexV3/features/Swap';
@@ -67,7 +66,7 @@ const PrivacyApps = () => {
       },
       {
         privacyAppId: KEYS_PLATFORMS_SUPPORTED.uni,
-        icon: <UniIcon />,
+        icon: <UniIcon2 />,
         headerTitle: 'pUniswap',
         headerSub: 'Private Uniswap',
         groupActions: [
